telegram-bot/bot-parser/requirements.txt
pytelegrambotapi
lxml
beautifulsoup4
requests
python-decouple